backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
sendgrid==6.10.0
requests==2.31.0
email-validator>=2.1.0
elasticsearch==8.11.0
boto3==1.34.0
pusher==3.3.2
sentry-sdk[fastapi]>=1.38.0
psutil==5.9.6
stripe==7.7.0
pyotp==2.9.0
qrcode[pil]==7.4.2
bleach==6.1.0
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
slowapi>=0.1.9